refactor(computer-use): put Windows capture with the rest of the Windows leaf
The helper binary carried two personalities: the RPC service the product spawns,
and a capture CLI nothing calls. The WGC and Direct3D code lived in the binary's
inline module, and the server imported it back out -- `super::super::super::` from
a platform leaf, three levels up into the executable's crate root. The reusable
layer depended on its own entry point, so it could not be built without it.

macOS was already the shape this should be: capture in platform_macos, a main of
thirteen lines. Windows now matches. The capture core moves to
platform_windows::wgc, beside the UIA and SendInput leaves it belongs with, and
window geometry joins window.rs where the other window queries are. Both entry
points collapse into one, so the cfg asymmetry and the comment explaining why
macOS had to attach the server differently both go.

The capture CLI is deleted: capture-window and capture-monitor appear nowhere
outside that file, and the product spawns only `serve --pipe`. Eleven items go
with it, and moving the rest exposed four more -- CaptureInfo carried ok, method,
hwnd and path, which existed to be printed as the CLI's JSON line and are read by
nobody.

The bodies were carved by line range rather than retyped, so the D3D11 and WinRT
code is byte-identical. Sixteen items kept, all present; eleven removed, none
lingering; no `super::super::super::` remains anywhere in the server. Windows
checks with no warnings and 21 tests pass.
