+++ console/src-tauri/src/computer_use_server/platform_windows/capture.rs
@@ -12,7 +12,7 @@ use windows::Graphics::Imaging::{
 };
 use windows::Storage::Streams::{DataReader, InMemoryRandomAccessStream};
 
-use super::super::super::{capture_window, CaptureArgs};
+use super::wgc::{capture_window, CaptureArgs};
 use super::super::state::{
     next_id, AccessibilitySnapshot, ServerState, Snapshot, WindowInfo, BMP_HEADER_BYTES,
     SCREENSHOT_JPEG_QUALITY, SCREENSHOT_MAX_EDGE,
