+++ console/src-tauri/src/computer_use_server/platform_windows/mod.rs
@@ -9,6 +9,7 @@
 mod capture;
 mod input;
 mod uia;
+mod wgc;
 mod window;
 
 pub(super) use capture::observe_window;
